requirements.txt
pyserial
pyserial-asyncio